use lstinputlisting for code snippets for consistency [skip ci]
