+++ .clang-format
@@ -1,0 +1,1 @@
+../graph-v2/.clang-format
